fix: Refactor server to conditionally import Vite
This commit fixes a production crash (ERR_MODULE_NOT_FOUND) that occurred because the server was trying to import the `vite` package in a production environment where dev dependencies are not installed.

- Refactored the server code by separating the Vite-powered development server logic into `server/dev.ts`.
- The production static file serving logic is now in `server/static.ts`.
- The main server entry point (`server/index.ts`) now uses a dynamic `import()` to load the development server only when `NODE_ENV` is 'development'. This ensures the `vite` package is never imported in production.

Additionally, this commit includes fixes for TypeScript errors that were identified during the build process:
- Corrected the `esbuild` command in `package.json` to use `--outfile` to create a single server bundle.
- Added type assertions in `ResearchAreas.tsx` to resolve implicit 'any' type errors.
- Used `parseInt` for the server port in `server/index.ts` to ensure it's a number.
