+++ package.json
@@ -5,7 +5,7 @@
     "license": "MIT",
     "scripts": {
         "dev": "tsx server/index.ts",
-        "build": "vite build && esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outdir=dist",
+        "build": "vite build && esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/index.js",
         "start": "NODE_ENV=production node dist/index.js",
         "check": "tsc",
         "db:push": "drizzle-kit push"
